added search locus page, create locus list

+++ src/org/irri/iric/portal/domain/VariantTableArray.java
@@ -1,9 +1,11 @@
 package org.irri.iric.portal.domain;
 
+import java.math.BigDecimal;
+
 public interface VariantTableArray extends VariantTable {
 
 
-	public Double[] getPosition();
+	public BigDecimal[] getPosition();
 
 	public String[] getReference();
